feat: footer
