+++ src/app/layout.tsx
@@ -21,7 +21,7 @@ export default function RootLayout({
       <body>
         {/* <Navbar /> */}
         {children}
-        {/* <Footer /> */}
+        <Footer />
       </body>
     </html>
   );
