+++ src/components/Footer.tsx
@@ -1,95 +1,34 @@
-"use client";
-
 import Image from "next/image";
 import brandLogo from "@/assets/BrandLogo.png";
 
-import {
-  faFacebook,
-  faGithub,
-  faLinkedin,
-  faInstagram,
-} from "@fortawesome/free-brands-svg-icons";
+import { socials } from "@/utils/socials";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 
 export default function Footer() {
   return (
-    <footer className="footer py-5">
-      <div className="d-flex flex-sm-row flex-column justify-content-center align-items-center">
-        <div className="col-4 col-md-3 my-3 order-2 order-sm-1">
-          <Image
-            src={brandLogo}
-            width="100"
-            alt="OSDG"
-            className="img-fluid mb-3"
-          />
-          <p className="footer-text m-0">&copy; 2024</p>
-          <p className="footer-text m-0">Open Source Developers Group</p>
-          <p className="footer-text m-0">
-            International Institute of Information
-          </p>
-          <p className="footer-text m-0">Technology, Hyderabad</p>
-        </div>
-        <div className="d-flex flex-column justify-content-center order-1 order-sm-2 col-3 col-md-3 my-3">
-          <div className="d-flex flex-row  justify-content-center">
-            <h2
-              className="social-link mx-2"
-              onClick={() => {
-                window.open(
-                  "https://www.facebook.com/groups/185567594878116",
-                  "_blank"
-                );
-              }}
-            >
-              <FontAwesomeIcon icon={faFacebook} />
-            </h2>
-            <h2
-              className="social-link mx-2"
-              onClick={() => {
-                window.open("https://github.com/OSDG-IIITH", "_blank");
-              }}
-            >
-              <FontAwesomeIcon icon={faGithub} />
-            </h2>
-            <h2
-              className="social-link mx-2"
-              onClick={() => {
-                window.open(
-                  "https://www.linkedin.com/company/74330374",
-                  "_blank"
-                );
-              }}
-            >
-              <FontAwesomeIcon icon={faLinkedin} />
-            </h2>
-            <h2
-              className="social-link mx-2"
-              onClick={() => {
-                window.open("https://www.instagram.com/osdg.iiith", "_blank");
-              }}
-            >
-              <FontAwesomeIcon icon={faInstagram} />
-            </h2>
-          </div>
-        </div>
-        <div className="col-2 d-flex my-3 text-end order-sm-3 d-none d-sm-block">
-          <p className="m-0">Reach Us</p>
-          <a
-            className="footer-text my-1 d-block"
-            href="mailto:osdg@students.iiit.ac.in"
-          >
-            osdg@students.iiit.ac.in
-          </a>
-        </div>
-        <div className="d-block d-sm-none text-center pb-3">
-          <p className="">Reach Us</p>
-          <a
-            className="footer-text my-1 d-block"
-            href="mailto:osdg@students.iiit.ac.in"
-          >
-            osdg@students.iiit.ac.in
-          </a>
-        </div>
+    <div className="flex justify-around px-32 py-12 bg-slate-700 text-white">
+      <div className="text-xs h-auto">
+        <Image src={brandLogo} alt="OSDG" className="w-32 mb-6" />
+        <p>&copy; 2024</p>
+        <p>Open Source Developers Group</p>
+        <p>International Institute of Information</p>
+        <p>Technology, Hyderabad</p>
       </div>
-    </footer>
+      <div className="flex h-auto justify-center items-center">
+        {socials.map(([link, icon], idx) => (
+          <a
+            href={link}
+            key={idx}
+            className="mx-2 opacity-60 hover:opacity-100 w-8 h-8"
+          >
+            <FontAwesomeIcon icon={icon} />
+          </a>
+        ))}
+      </div>
+      <div className="h-auto flex flex-col justify-center items-end">
+        <p className="text-l font-bold">Reach Us</p>
+        <a  className="text-xs" href="mailto:osdg@students.iiit.ac.in">osdg@students.iiit.ac.in</a>
+      </div>
+    </div>
   );
 }
