+++ src/utils/socials.ts
@@ -1,0 +1,16 @@
+import { IconDefinition } from "@fortawesome/free-brands-svg-icons";
+import {
+  faFacebook,
+  faGithub,
+  faLinkedin,
+  faInstagram,
+} from "@fortawesome/free-brands-svg-icons";
+
+const socials: [string, IconDefinition][] = [
+  ["https://www.facebook.com/groups/185567594878116", faFacebook],
+  ["https://github.com/OSDG-IIITH", faGithub],
+  ["https://www.linkedin.com/company/74330374", faLinkedin],
+  ["https://www.instagram.com/osdg.iiith", faInstagram],
+];
+
+export { socials };
